Add the option to rewrite the root of files
This change allows the URI of the files that Debug Bar needs to be re-written. When used with symlinks, .htaccess, creating routes to pass through the files, or other methods this can be used to solve the 404 problem when files aren't where Debug Bar expects them to be. This also provides a way to solve problems that arise from using Debug Bar with the base-path middleware.

As an example my project lives in a subfolder of my domain AND I have my vendor directory above my public folder. By rewriting the URI path for the Debug Bar files with ``(new Middlewares\Debugbar())->fileURI("/MyProjectsSubDirectory/maximebf/debugbar/")`` I am then able to create corresponding routes that link back to the files in the "../src/vendor/maximebf/debugbar/src/DebugBar/Resources/" location that isn't publicly accessible.